Add files via upload
This is my first project for my portfolio. It is code done on SQL in order to discover more about data and be able to extract whatever is needed in order to visualize it afterwards on Tableau. It was very interesting and it was also very enlightening not only because of the coding process, the edit and the extraction but also because it was a very interesting data set to begin with. So feel free to take a look around and you can also follow the link to my Tableau Dashboard where the visualization part is located. Thank you for reading!